Resolve naming inconsistency and associated equivalent class complexity
This patch applies global find and replaces, using these patterns:

    s/case-core:Trace/uco-observable:CyberItem/g
    s/case-core:hasPropertyBundle/uco-core:facets/g

This patch also removes the obviated case-core namespace prefix in
context dictionaries, replacing it with uco-core and uco-observable
where needed.

This patch was confirmed to not cause any changes in the
CASE-Examples-QC migration status lists, and to undo changes caused by
the immediately prior patch.

References:
* [ONT-231] (CP-10) Definition of Trace from CASE prototype not
  preserved
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>