Migrate remaining terms to CASE 0.2.0.
Thanks to Nikolaos, namespaces were populated for the CASE prototype
terms.  However, while Nikolaos was working, CASE 0.2.0 was released.
This patch steps through the namespaces and aligns them with CASE 0.2.0
and UCO 0.4.0, using the AC-118 branch on the CASE-Examples-QC
repository.

This patch closes AC-80, and the merge will close AC-118.

References:
* [AC-80] Forensic Lifecycle example migration
* [AC-118] Update CASE-Examples repository to point CASE submodule to
  0.2.0 release
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC/

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>