Build "Serialized" ontology for UCO Validator codebase
The Validator code base expects a single argument to specify the
ontology for its "serialize" command: Either read a single ontology
file, or "a directory containing turtle files."  For the latter argument
form, the current version of that code base does not look in
sub-directories.  Hence, this patch adds soft-links to gather all of the
ontology files for CASE and UCO into a single directory with no deeper
walking required.

To confirm the serialization file builds, this patch also starts a
"check" target, using the Validator code base's "describe" function to
confirm the pickle file is as the Validator code base expects.  And,
since a long list of link files was created, a link-liveness test is
included as well in dependencies/CASE-0.3.0.

References:
* [AC-137] Add per-CASE-version validation to CASE-Examples

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>