Address errant location association
Recent changes caused an odd-looking drop from the CASE-Examples-QC
all-concepts-used list.  uco-observable:Location was no longer being
used.  One of the location references, errant for similar reasons to the
annotation attached to the first one, was still intact.

With the discussion on the Locations MVP proposal, there is now a way
forward for representing this desired association.  This patch fixes the
error while we draft that change.

References:
* [CASE-Examples-QC] https://github.com/ajnelson-nist/CASE-Examples-QC
* [ONT-186] (CP-30) Topic - Location