(DO NOT MERGE) Remove Validator from build chain for CI demo
This patch removes the Validator due to an upstream build issue.

This patch is not intended to be merged, as the effect of the build
issue is expected to be temporary.

This patch also caught an errant recursion call in the top
`make check`, expected to be fixed in another patch.