Bump validator dependency to support descending 'make check'
The top-level `check` target wasn't recursively calling `check` in the
illustrations directory.  The Validator needed to have its
import-reference bumped to resolve an error fixed since the last time it
was tracked (in Validator PR 13).

With this patch, `make check` recursively runs `check` in the
illustrations directory without returning an error state.  The Validator
reports data issues, but does not exit in an error state, so CI should
continue to run without hiccups.

References:
* https://github.com/ucoProject/UCO-Utility-Pre-0.7.0-Validator/pull/13

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>