Remove pre-CASE-0.5.0 validation process
The UCO pre-0.7.0 Validator was written to provide a stopgap validation
need while UCO and CASE transitioned to SHACL.  With the release of
CASE 0.5.0 and `case_utils` 0.3.0, the tool `case_validate` now provides
that need.  The CASE 0.3.0 and 0.4.0 validation process is removed in
this patch, so the community can focus its efforts on validation results
against the current ontology.

References:
* [AC-207] Release CASE-Utilities-Python 0.3.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>