Address some SHACL validation errors
This patch revises the JSON-LD files to address issues reported in SHACL
validation against CASE 0.5.0/UCO 0.7.0.  This was done to ease the
review complexity for the UCO 0.8.0 release.

Objects that needed to be created anew followed the actual-UUID practice
prescribed in AC-148.

This is a follow-on to a prior patch in the same spirit.

A follow-on patch will regenerate the Make-managed files.

References:
* [UCO OC-113] UCO Version 0.8.0
* [ONT-463] Release CASE 0.6.0

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>