Remove all vocabulary datatype associations for kindOfRelationship
This reduction to `xsd:string` is to accommodate what was found to be
the best possible resolution for SHACL enforcement of semi-open
vocabularies, at the time of UCO 0.8.0's release.  Some future
proposal(s) will restore association of `kindOfRelationship` with
controlled value-sets.

A follow-on patch will regenerate Make-managed files.

References:
* [OC-12] (CP-100) UCO's idea of "Open vocabulary" does not agree with
  its implementation with owl:oneOf

Requested-by: Sean Barnum <sbarnum@mitre.org>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>