Add HermiT download and extraction with TOFU signature
The HermiT tool is one of the tools listed among the OWL 2 test suite
evaluation [1].  It is being tracked in this repository to assist with
reviewing some ontology design pattern implementations.

The URL recorded in `lib/Makefile` is the HTTPS redirection URL when
retrieving this URL:

    http://www.hermit-reasoner.com/download/1.3.8/HermiT.zip

The .sha512 is a signature stored in a trust-on-first-use protocol.
A second download source could not be found (archive.org was searched)
to help verify the file contents had not changed since original posting.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

References:
* [ONT-434] (CP-33) CASE needs to provide an inference object
* [1] https://www.w3.org/2007/OWL/wiki/index.php?title=Test_Suite_Status

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>