Add configured tool illustration
AJN: This is a transcription of an example Sean Barnum drafted in
response to Issue 405 (comment cited in References).  I've annotated
this JSON file as authored by him in the Git history, only changing
according to effects from `pre-commit` review.  The file is posted here
with Sean's permission, for the sake of SHACL review and inspection of
new concepts under draft.

References:
* https://github.com/ucoProject/UCO/issues/405#issuecomment-1180785199

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>