Update example assuming application of sh:xone shape selector
This patch, and the follow-on patch regenerating the validation results,
demonstrates a difficult-to-diagnose issue with the overly explicit,
inlined-in-Class `sh:PropertyShape` pattern used in UCO.

This patch adopts the current implementation for Issue 405, where a
multi-Constraint-bearing `sh:PropertyShape` is inlined in the `sh:xone`
branches.  This patch does not correct an error, where this node is not
defined:

`kb:File-0fedea11-3c8b-400c-9cdb-593edc146194`

Because the `sh:xone` shapes each define a conjunction of requirements
that includes the `sh:class` constraint on the referenced object, all
three of the `sh:xone` shapes fail.  The `itemObject` property shape
requires exactly one `itemObject`, AND that the `itemObject` be a
`UcoObject`.  Failing either condition fails the whole
`sh:PropertyShape`.

The correct solution here is for the `sh:xone` to minimally specify the
property shape's constraints within the `sh:xone`.  The other
constraints should either go into UCO's pattern of explicit inlining
onto the property, or into a class-independent `sh:PropertyShape` using
`sh:targetSubjectsOf`.

I'll add a patch to the UCO 405 branch to follow UCO's current pattern.

References:
* https://github.com/ucoProject/UCO/issues/405

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>