Revert UUID pattern for one example cell site
The node renamed in `cell_site.json` was intended to not follow the UUID
pattern of Issue 430, and instead demonstrate an alternative mechanism
for building a knowledge base of cell sites.  This one node is extempted
from the UUID rename until community feedback provides an argument to
change it.

No effects were observed on Make-managed files from this reversion.

References:
* https://github.com/ucoProject/UCO/issues/430

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>