Prevent default typing as xsd:double
JSON-LD has a default conversion of non-integer literals to
`xsd:double`, per Section B.1.3 of the JSON-LD 1.1 specification.

What does not appear to be explained in the specification---at least, in
that section---is what to do when there is an `@type` annotation
accompanying a non-string `@value`.  RDFLib, at least as of 6.2.0, does
not use the `@type` annotation.  This makes the pySHACL output look
quite mysterious.

This patch converts the `@value` to a string so `@type` will be
respected.

A follow-on patch will regenerate Make-managed files.

References:
* https://w3c.github.io/json-ld-syntax/#conversion-of-native-data-types

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>