Consolidate non-standard concepts into one drafting namespace
This uses the `drafting:` prefix used on the CASE website and in other
examples making use of `drafting.ttl` files.

The unprefixed concepts have, to date, appeared like typos, relying on
the `@vocabulary` JSON-LD construct to preserve the concepts'
functionality as graph data.  This practice has also hidden the concepts
from maintainers' attention, so this patch reduces the usage of the
`@vocabulary` support.  This will have the additional benefit of
reducing the number of randomly-assigned namespace prefixes in the SHACL
validation results.

A follow-on patch will regenerate Make-managed files.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>