Have Method be a subclass of Pattern instead of ActionLifecycle
This is a guess, taken for a reason raised by SHACL validation:
An instance of ActionLifecycle requires a property `action:phase` be
assigned, and it seems this will frequently be unavailable for methods
that are stub references to large libraries of methods.

Note that, at the time of this writing, ActionLifecycle and Pattern both
need further demonstration among CASE or UCO examples.

A follow-on patch will regenerate Make-managed files.

References:
* [ONT-434] (CP-33) CASE needs to provide an inference object

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>