Transcribe JSON from Issue 400
AJN: This is me transcribing what Sean had illustrated in UCO Issue 400.
As I recall, he and Eoghan had drafted this together.  Sean noted in an
Ontology Committees meeting this could be transcribed to CASE-Examples
for SHACL review and other review.

As part of transcribing this, `pre-commit` adjusted indentation, and I
fixed JSON syntax (a few errant comments).

References:
* https://github.com/ucoProject/UCO/issues/400

Co-authored-by: Eoghan Casey <109762296+eoghanscasey@users.noreply.github.com>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>