Add proof of concept script and framework for Relationship illustration
This also sets up README sections, and in some cases stub README files,
to show the availability of the Relationship renders.

A follow-on patch will regenerate Make-managed files.

Disclaimer:
Participation by NIST in the creation of the documentation of mentioned
software is not intended to imply a recommendation or endorsement by the
National Institute of Standards and Technology, nor is it intended to
imply that any specific software is necessarily the best available for
the purpose.

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>