Move name of message threads
`uco-observable:displayName` is not associated with
`uco-observable:MessageThreadFacet` by a SHACL property shape, at the
time of UCO 1.3.0.

Nothing encoded in UCO disallows the way it was used in `message.json`,
but the data stored in that property is being removed to follow the
suggestions from the SHACL shapes.

Reported in CASE-Mapping-Python PR 38.

No effects were observed on Make-managed files.

References:
* https://github.com/casework/CASE-Mapping-Python/pull/38

Reported-by: Fabrizio Turchi <fabrizio.turchi@cnr.it>
Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>