Align tool identifiers
A UUID was shared between two non-matching IRIs of a tool.  This patch
updates the IRI and types to use the analysis example's IRI, and aligns
the creating organization.

Because the configured tool was a specialization of `uco-tool:Tool`,
this patch carries the specialization forward by adding an extra type to
the configured tool node.

The duplicated UUID was found while reviewing CASE-Examples-QC PR 54.

No effects were observed on Make-managed files.

References:
* https://github.com/ajnelson-nist/CASE-Examples-QC/pull/54

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>