Align organization identifier
A UUID was shared between two non-matching IRIs of an organization
between examples.  This was in part due to a prefix-string using a
different prefix-IRI in the spear phishing example.  This patch defers
to the example IRI used elsewhere across CASE examples.

The duplicated UUID was found while reviewing CASE-Examples-QC PR 54.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ajnelson-nist/CASE-Examples-QC/pull/54

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>