Align software identifier
A UUID was shared between two non-matching IRIs of a node typed as
Software in one example and a File in another example.  This patch
changes the spelling in both examples to share the node.

The duplicated UUID was found while reviewing CASE-Examples-QC PR 54.

A follow-on patch will regenerate Make-managed files.

References:
* https://github.com/ajnelson-nist/CASE-Examples-QC/pull/54

Signed-off-by: Alex Nelson <alexander.nelson@nist.gov>